LayoutTests/http/tests/navigation/page-cache-requestAnimationFrame-expected.txt

Tests that requestAnimationFrame() does not break when restoring from PageCache upon navigating back cross-origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS requestAnimationFrame is running after restoring from PageCache.
PASS successfullyParsed is true

TEST COMPLETE
